// src/app/models/customer/customer.model.ts
export default class Customer {
  id: string = '';
  branch: string = '';
  street: string = '';
  houseNumber: string = '';
  postalCode: string = '';
  city: string = '';
  owner: string = '';
  email: string = '';
  telephone: string = '';
  fax: string = '';
  homePage: string = '';
  notations: string = '';
  createdOn: string = '';
  createdBy: string = '';
}